# internal/dashboard/static/openapi.yaml
openapi: 3.0.3
info:
  title: IaaS Platform API
  description: |
    A Go-based cloud control plane and billing platform. Multi-tenant
    organizations, compute resource lifecycle management, API key
    authentication, rate limiting, quotas, and usage-based billing.
    The compute backend is a simulation; the control-plane API is real.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Control-plane API
  - url: http://localhost:8080
    description: Health probes (served on the server root)
...
paths:
...
  /auth/signup:
    post:
      tags: [auth]
      summary: Create an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Email already taken
        '413':
          $ref: '#/components/responses/RequestEntityTooLarge'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /auth/login:
    post:
      tags: [auth]
      summary: Log in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Invalid credentials
  /auth/forgot-password:
    post:
      tags: [auth]
      summary: Request a password reset email
      description: Always reports success for valid requests so callers cannot enumerate which emails have accounts.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '200':
          description: Reset email sent (or would have been sent)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForgotPasswordResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /auth/reset-password:
    post:
      tags: [auth]
      summary: Reset a password with a one-time token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForgotPasswordResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /me:
    get:
      tags: [auth]
      summary: Get the current user
      security:
        - bearerAuth: []
        - apiKey: []
      responses:
        '200':
          description: The authenticated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /orgs:
    post:
      tags: [organizations]
      summary: Create an organization
      security:
        - bearerAuth: []
        - apiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrgRequest'
      responses:
        '201':
          description: Organization created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Slug already taken
    get:
      tags: [organizations]
      summary: List organizations for the current user
      security:
        - bearerAuth: []
        - apiKey: []
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Organizations
          headers:
            X-Total-Count:
              $ref: '#/components/headers/XTotalCount'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Organization'
  /orgs/{orgID}:
    parameters:
      - $ref: '#/components/parameters/OrgID'
    get:
      tags: [organizations]
      summary: Get an organization
      security:
        - bearerAuth: []
        - apiKey: []
      responses:
        '200':
          description: Organization
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  headers:
    XTotalCount:
      description: Total number of items available, not just the current page
      schema:
        type: integer
        format: int64
    X-RateLimit-Limit:
      description: Maximum number of requests allowed per burst window
      schema:
        type: integer
    X-RateLimit-Remaining:
      description: Number of requests remaining in the current window
      schema:
        type: integer
    X-RateLimit-Reset:
      description: Seconds until the rate-limit window resets (only on 429)
      schema:
        type: integer
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    Limit:
      name: limit
      in: query
      required: false
      description: Maximum number of items to return (default 50, max 100)
      schema:
        type: integer
        minimum: 1
        maximum: 100
    Offset:
      name: offset
      in: query
      required: false
      description: Number of items to skip before returning results
      schema:
        type: integer
        minimum: 0
    OrgID:
      name: orgID
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/X-RateLimit-Limit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/X-RateLimit-Remaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/X-RateLimit-Reset'
        Retry-After:
          description: Seconds to wait before retrying
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RateLimitError'
    RequestEntityTooLarge:
      description: Request body exceeds size limit
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: invalid request
    RateLimitError:
      type: object
      required: [error, retry_after]
      properties:
        error:
          type: string
          example: rate limit exceeded
        retry_after:
          type: integer
          description: Seconds to wait before retrying
          example: 1
    SignupRequest:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
        name:
          type: string
        organization:
          type: string
        org_slug:
          type: string
          description: Slug of an existing organization to request access to. The organization admin must accept before access is granted.
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        api_key:
          type: string
          description: Raw API key. Returned only once, on signup; the server stores a SHA-256 hash.
        org_join_pending:
          type: boolean
          description: True when the account requested access to an organization and is awaiting admin approval.
        user:
          $ref: '#/components/schemas/User'
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    ResetPasswordRequest:
      type: object
      required: [token, new_password]
      properties:
        token:
          type: string
          description: Single-use reset token from the password reset email.
          minLength: 20
        new_password:
          type: string
          format: password
          minLength: 8
    ForgotPasswordResponse:
      type: object
      properties:
        status:
          type: string
          example: sent
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        name:
          type: string
        role:
          type: string
        organization:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateOrgRequest:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 64
        slug:
          type: string
          minLength: 3
          maxLength: 32
    Organization:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        slug:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...